requirements.txt
pycryptodome==3.4
boto3=1.4.0